refactor(kernel): split local vcs command mapping
